+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `simulation` checkout query.
+
+### Deprecated
+- `shipping` query.
 
 ## [2.70.4] - 2019-05-06
 ### Added